+++ android/app/src/main/java/com/hogent/android/network/services/ContractApiService.kt
@@ -14,7 +14,7 @@ private val retrofit = Config.createRetrofit(API)
 interface ContractApiService{
 
     @GET("{id}")
-    suspend fun getById(@Path("id") id: Long): Contract?
+    suspend fun getById(@Path("id") id: Int): Contract?
 
     @POST
     suspend fun createContract(@Body contract: Contract): Contract
